accounts/api/views.py
from rest_framework import permissions
from rest_framework.response import Response
from rest_framework.views import APIView

from .serializers import UserReadSerializer


class UserDetailAPIView(APIView):
    permission_classes = (permissions.IsAuthenticated,)

    def get(self, request, *args, **kwargs):
        serializer = UserReadSerializer(request.user)
        return Response(data=serializer.data)
