requirements/compiled-dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/compiled-dev.txt --resolver=backtracking requirements/dev.txt
#
asgiref==3.5.2
    # via django
asttokens==2.1.0
    # via stack-data
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via ipython
black==22.10.0
    # via -r requirements/lint.txt
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via black
coverage==6.5.0
    # via -r requirements/test.txt
crispy-bootstrap5==0.6
    # via -r requirements/base.txt
cryptography==38.0.3
    # via pyjwt
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via python3-openid
distlib==0.3.6
    # via virtualenv
django==4.1.1
    # via
    #   -r requirements/base.txt
    #   crispy-bootstrap5
    #   django-allauth
    #   django-debug-toolbar
    #   django-extensions
    #   djangorestframework
django-allauth==0.51.0
    # via -r requirements/base.txt
django-crispy-forms==1.14.0
    # via
    #   -r requirements/base.txt
    #   crispy-bootstrap5
django-debug-toolbar==3.2.4
    # via -r requirements/dev.txt
django-environ==0.9.0
    # via -r requirements/base.txt
django-extensions==3.2.1
    # via -r requirements/base.txt
djangorestframework==3.14.0
    # via -r requirements/base.txt
exceptiongroup==1.0.4
    # via pytest
executing==1.2.0
    # via stack-data
filelock==3.8.0
    # via virtualenv
flake8==5.0.4
    # via
    #   -r requirements/lint.txt
    #   flake8-print
flake8-print==5.0.0
    # via -r requirements/lint.txt
gunicorn==20.1.0
    # via -r requirements/production.txt
identify==2.5.9
    # via pre-commit
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
ipython==8.6.0
    # via -r requirements/dev.txt
isort==5.10.1
    # via -r requirements/lint.txt
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.7.0
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==21.3
    # via pytest
parso==0.8.3
    # via jedi
pathspec==0.10.2
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.4
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r requirements/lint.txt
prompt-toolkit==3.0.32
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements/base.txt
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via
    #   flake8
    #   flake8-print
pycparser==2.21
    # via cffi
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via ipython
pyjwt[crypto]==2.6.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements/test.txt
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via asttokens
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.1
    # via ipython
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   pytest
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.12
    # via requests
virtualenv==20.16.7
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==6.2.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
